[DEV-4075] combine spending_test_fixtures files
