+++ usaspending_api/search/tests/integration/test_spending_by_country.py
@@ -7,16 +7,13 @@ from rest_framework import status
 from usaspending_api.common.experimental_api_flags import ELASTICSEARCH_HEADER_VALUE, EXPERIMENTAL_API_HEADER
 from usaspending_api.common.helpers.generic_helper import get_time_period_message
 from usaspending_api.search.tests.data.search_filters_test_data import non_legacy_filters
-from usaspending_api.search.tests.integration.spending_test_fixtures import (
+from usaspending_api.search.tests.integration.spending_by_category.spending_test_fixtures import (
     setup_basic_awards_and_transactions,
-    setup_country_data,
-    setup_basic_agency_tree,
+    setup_country_data
 )
-
 
 @pytest.fixture
 def country_test_data(db):
-    setup_basic_agency_tree()
     setup_basic_awards_and_transactions()
     setup_country_data()
